ServerMain without EngineMaintenanceJob
